fix(UI): BeerDrawer as a separate component
